requirements.txt
dash==1.17.0
pandas==1.1.3
plotly==4.11.0
gunicorn
google-api-python-client
google-auth-oauthlib
geopy==2.0.0
opencage==1.2.1